Refactor with service file
