+++ ChordKTV/Dtos/HandwritingCanvas/HandwritingCanvasDto.cs
@@ -2,5 +2,6 @@ namespace ChordKTV.Dtos;
 
 public record HandwritingCanvasDto(
     string Image,
-    LanguageCode Language
+    LanguageCode Language,
+    string ExpectedMatch
 );
