+++ ChordKTV/Services/Api/IHandwritingService.cs
@@ -1,0 +1,8 @@
+using ChordKTV.Dtos;
+
+namespace ChordKTV.Services.Api;
+
+public interface IHandwritingService
+{
+    public Task<bool> RecognizeTextAsync(HandwritingCanvasDto image);
+}
